refactor(backup): add top-level service barrel
Per naming-conventions §6.4, add src/main/services/backup/index.ts as a
pure re-export barrel declaring the directory's encapsulation boundary:
serviceRegistry now imports BackupService from '@main/services/backup'
instead of the deep BackupService path. Route BackupService's own
contributorManager import through the contributors sub-barrel and
normalize its export order so both barrels stay pure re-exports.
