+++ src/main/core/application/serviceRegistry.ts
@@ -26,7 +26,7 @@ import { IpcApiService } from '@main/ipc/IpcApiService'
 import { AnalyticsService } from '@main/services/AnalyticsService'
 import { AppMenuService } from '@main/services/AppMenuService'
 import { AppUpdaterService } from '@main/services/AppUpdaterService'
-import { BackupService } from '@main/services/backup/BackupService'
+import { BackupService } from '@main/services/backup'
 import { BinaryManager } from '@main/services/BinaryManager'
 import { CodeCliService } from '@main/services/codeCli'
 import { CommandService } from '@main/services/CommandService'
